+++ x/group/CHANGELOG.md
@@ -33,3 +33,5 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Features
 
 ### API Breaking Changes
+
+* [#19410](https://github.com/cosmos/cosmos-sdk/pull/19410) Migrate to Store Service.